stack and queue

+++ Data structures/List.h
@@ -11,7 +11,7 @@ public:
 	virtual void pushFront(const T& el) = 0;
 	virtual void insertAt(const T& el, int index) = 0;
 
-	virtual T getAt(int index) = 0;
+	virtual T getAt(size_t index) = 0;
 
 	virtual T popBack() = 0;
 	virtual T popFront() = 0;
